TINY-6920: Moved test, slight improvement.
